# Add renderer low latency mode