--- a/doc/classes/RenderingServer.xml
+++ b/doc/classes/RenderingServer.xml
@@ -1592,6 +1592,13 @@
 				Tries to free an object in the RenderingServer. To avoid memory leaks, this should be called after using an object as memory management does not occur automatically when using RenderingServer directly.
 			</description>
 		</method>
+		<method name="get_cpu_gpu_sync_mode" qualifiers="const">
+			<return type="int" enum="RenderingServer.CpuGpuSyncMode" />
+			<description>
+				Gets the CPU/GPU synchronization mode for the renderer. See [enum CpuGpuSyncMode] for options.
+				[b]Note:[/b] This setting should be used with care, as the sequential low latency mode prevents the CPU and GPU from running in parallel, which may adversely affect performance by as much as 30-50%. For simple scenes with capped FPS, this may not be noticeable. In other scenarios, the performance impact is more severe. An FPS drop, e.g. from 1000 to 500 FPS, or from 90 to 45 FPS, is an expected outcome from using sequential mode. Therefore, it is recommended that this option is exposed in your game's settings menu, so that users can select the best mode based on their preferences and hardware.
+			</description>
+		</method>
 		<method name="get_current_rendering_driver_name" qualifiers="const">
 			<return type="String" />
 			<description>
@@ -3415,6 +3422,14 @@
 			<param index="3" name="use_filter" type="bool" default="true" />
 			<description>
 				Sets a boot image. The color defines the background color. If [param scale] is [code]true[/code], the image will be scaled to fit the screen size. If [param use_filter] is [code]true[/code], the image will be scaled with linear interpolation. If [param use_filter] is [code]false[/code], the image will be scaled with nearest-neighbor interpolation.
+			</description>
+		</method>
+		<method name="set_cpu_gpu_sync_mode">
+			<return type="void" />
+			<param index="0" name="sync_mode" type="int" enum="RenderingServer.CpuGpuSyncMode" />
+			<description>
+				Sets the CPU/GPU synchronization mode for the renderer. See [enum CpuGpuSyncMode] for options. Equivalent to [member ProjectSettings.rendering/driver/synchronization/cpu_gpu_sync].
+				[b]Note:[/b] This setting should be used with care, as the sequential low latency mode prevents the CPU and GPU from running in parallel, which may adversely affect performance by as much as 30-50%. For simple scenes with capped FPS, this may not be noticeable. In other scenarios, the performance impact is more severe. An FPS drop, e.g. from 1000 to 500 FPS, or from 90 to 45 FPS, is an expected outcome from using sequential mode. Therefore, it is recommended that this option is exposed in your game's settings menu, so that users can select the best mode based on their preferences and hardware.
 			</description>
 		</method>
 		<method name="set_debug_generate_wireframes">
@@ -5878,6 +5893,12 @@
 		</constant>
 		<constant name="GLOBAL_VAR_TYPE_MAX" value="29" enum="GlobalShaderParameterType">
 			Represents the size of the [enum GlobalShaderParameterType] enum.
+		</constant>
+		<constant name="CPU_GPU_SYNC_PARALLEL" value="0" enum="CpuGpuSyncMode">
+			Tells the renderer to prioritize higher framerate by allowing the CPU to queue up additional frames before they're rendered by the GPU. This allows the CPU and GPU to work in tandem, improving the framerate and framepacing in complex scenes at the expense of input latency. This default setting is suitable for most 3D applications, especially on mobile and lower-performance desktop hardware.
+		</constant>
+		<constant name="CPU_GPU_SYNC_SEQUENTIAL" value="1" enum="CpuGpuSyncMode">
+			Tells the renderer to prioritize lower display latency by limiting how far the CPU is allowed to go ahead of the GPU when queueing frames. This can greatly help with input lag, at the cost of significantly reduced framerate in complex scenes. This setting is useful for games and applications with simple graphics where responsive input is important. Your results may vary based on platform, drivers, and scene contents.
 		</constant>
 		<constant name="RENDERING_INFO_TOTAL_OBJECTS_IN_FRAME" value="0" enum="RenderingInfo">
 			Number of objects rendered in the current 3D scene. This varies depending on camera position and rotation.