--- a/servers/rendering/dummy/storage/utilities.h
+++ b/servers/rendering/dummy/storage/utilities.h
@@ -75,6 +75,11 @@ public:
 	virtual uint64_t get_captured_timestamp_cpu_time(uint32_t p_index) const override { return 0; }
 	virtual String get_captured_timestamp_name(uint32_t p_index) const override { return String(); }
 
+	/* SYNCHRONIZATION */
+
+	virtual void set_cpu_gpu_sync_mode(RS::CpuGpuSyncMode p_sync_mode) override {}
+	virtual RS::CpuGpuSyncMode get_cpu_gpu_sync_mode() const override { return RS::CpuGpuSyncMode::CPU_GPU_SYNC_PARALLEL; }
+
 	/* MISC */
 
 	virtual void update_dirty_resources() override {}