… doc/classes/ProjectSettings.xml
… misc/dist/shell/_godot.zsh-completion
… misc/dist/shell/godot.bash-completion
… misc/dist/shell/godot.fish