Supersedes https://github.com/godotengine/godot/pull/94898

More or less implements https://github.com/godotengine/godot-proposals/issues/11200 . Instead of using `frame_queue_size = 1` though, this PR waits on the GPU using `_stall_for_previous_frames()` for RenderingDevice, and `glFinish()` for the Compatibility renderer. This means the swapchain and frame queue don't need to be recreated, which has two advantages:

1. Low latency mode can be enabled or disabled during runtime, meaning it can be easily added to the in-game settings menu.
2. Pipelining can be enabled/disabled on a per-frame basis.

The renderer latency mode can be adjusted via the `rendering/latency/low_latency_mode` project setting, as well as via `Engine::set_render_latency_mode()` and the `--renderer-latency` commandline argument. The two options are:

- `RENDER_LATENCY_PRIORITIZE_FRAMERATE` / `--renderer-latency high_framerate` - Tells the renderer to prioritize higher framerate by allowing the CPU to queue up additional frames before they're rendered by the GPU. This allows the CPU and GPU to work in tandem, improving the framerate and framepacing in complex scenes at the expense of input latency.
- `RENDER_LATENCY_PRIORITIZE_LOW_LATENCY` / `--renderer-latency low_latency` - Tells the renderer to prioritize lower display latency by limiting how far the CPU is allowed to go ahead of the GPU when queueing frames. This can greatly help with input lag, at the cost of significantly reduced framerate in complex scenes.

The enum names were chosen deliberately to highlight that a tradeoff is being made between high framerate and low latency. The documentation should also hopefully have ample warning so developers at least know what they're getting into with this setting.

In the future I'd like some sort of dynamic/automatic option which would switch pipelining on/off based on whether it would improve the framerate, inspired by the `auto_fps_auto_pipeline` Swappy mode available on Android. This would theoretically combine the benefits of both modes, with high performance under load and low latency where available.

The automatic approach has its disadvantages though, and I'd rather do more investigation before introducing an option that tries to be "smart" but ends up causing more problems. At the very least, it'd need to be aware of frame timings on both the CPU and GPU, to determine whether they would fit within one V-Sync period or if pipelining would be needed. AMD and nVidia's low latency SDKs/extensions might be more effective here than trying to come up with our own solution, but there's availability/licensing issues to contend with.

Right now this PR is mainly to test the waters and see if this would be satisfactory for a "low latency mode", or if more work is needed. This is just one of several improvements that need to be made to achieve the ultimate goal of *1 frame without pipelining*, or *2-3 frames with pipelining*. This PR just lets the user control whether pipelining is enabled or not. Eliminating unnecessary swapchain waits before drawing, and ensuring each renderer uses direct scanout (DXGI on Windows), are the two other main improvements to be made.

With this change, RenderingDevice saves 1 frame (at default settings) and OpenGL can save up to 2 frames (depending on drivers etc, YMMV). That means it's possible to achieve 1 frame of display latency in OpenGL, and 2 frames in Vulkan/D3D12, though at the cost of significantly reduced framerate in complex scenes. But if the game in question is a 2D game with simple graphics, there is very little risk from enabling this mode.