@Capewearer 2024-12-07
Sounds very great, could you explain how it would interact with double/triple buffering? Because it feels like turning off/on this exact feature.