@Capewearer 2024-12-07
> By double/triple buffering, do you mean the frame queue, or the swapchain? When enabled, this effectively forces a single-buffered frame queue, as if you set `frame_queue_size` to 1, but without actually using a smaller frame queue (size remains the same, just doesn't get filled). Right now there's no way to set `frame_queue_size` to 1 without modifying the engine, so with this change you can bring the latency one frame lower than you could otherwise.
> 
> The swapchain is unaffected, so if you want the lowest latency with standard (FIFO) V-Sync you'll need to set `swapchain_image_count` to 2.

Yes, that's what I really asked. Thanks for explanation!