@KeyboardDanni 2025-02-08
> ### Feedback
> * Should we use low-latency mode in the editor automatically? I think we can make it always be used in the 2D editor, script editor and AssetLib tabs. In the 3D editor, it would be enabled whenever the 3D editor viewport doesn't have focus (i.e. the mouse cursor isn't on the 3D editor viewport). This would avoid skewing performance measurements with the **View Frame Time** panel.
>   
>   * There could be an editor setting to override this behavior, with the values **Auto** (default), **Never**, **Always**.

I don't see much use case for having this in the editor, unless there's some weird crossover caused by the new game embedding feature.

Worth noting that the low latency gains here diminish significantly once you go beyond 120hz - this mode is primarily intended for 2D or lightweight 3D games running on 60hz displays. Also worth noting that nVidia Reflex and AMD Anti-Lag have driver-level knowledge that could minimize the FPS impact (since they're aware of the current CPU/GPU timings and can optimize for that), but this PR is at least a start.