@Calinou 2025-02-08
> I don't see much use case for having this in the editor, unless there's some weird crossover caused by the new game embedding feature.

We have a lot of issues related to input lag in the editor: https://github.com/godotengine/godot/issues/71795

Disabling V-Sync avoids these issues, but it introduces tearing (which is mostly an issue for the 2D/3D viewports). Also, regardless of V-sync, we can get further improvements to latency by reducing frame queuing.

In general, non-game applications are the area where I'd expect the low-latency mode to be most useful – and Godot is a non-game application made with Godot :slightly_smiling_face: