@KeyboardDanni 2025-02-08
Discussed with @Calinou and we'll probably handle low latency in the editor in a followup PR.