@KeyboardDanni 2025-02-27
This PR won't have any effect unless you manually enable the low latency mode in project settings:

![2025-02-27 12_54_31-rendering - Godot Contributors Chat](https://github.com/user-attachments/assets/0fcb252b-ac3b-4ffa-b4b0-d889f15f5fe2)

> Interestingly, I have consistently found that Compatibility has one extra frame of latency in fullscreen mode compared to windowed mode, even in 4.4rc1. Maybe I'll open an inssue for this...

This is driver behavior, I'm pretty sure. OpenGL lets drivers do additional frame queueing if they want, but are expected to flush that queue when `glFinish` is called (which this PR does with low latency mode active). Could also change the number of queued/pre-rendered frames in the driver settings. The driver-level low-latency setting might also work.

Driver settings won't have any effect in D3D12 or Vulkan though. With those APIs, frame queueing is handled by the client application, so the only way to disable that queueing is with this PR.