@allenwp 2025-02-27
> This PR won't have any effect unless you manually enable the low latency mode in project settings:
> 
> ![2025-02-27 12_54_31-rendering - Godot Contributors Chat](https://private-user-images.githubusercontent.com/34800072/417726070-0fcb252b-ac3b-4ffa-b4b0-d889f15f5fe2.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3NDA2Nzk3NzcsIm5iZiI6MTc0MDY3OTQ3NywicGF0aCI6Ii8zNDgwMDA3Mi80MTc3MjYwNzAtMGZjYjI1MmItYWMzYi00ZmZhLWI0YjAtZDg4OWYxNWY1ZmUyLnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNTAyMjclMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjUwMjI3VDE4MDQzN1omWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPTAxYmFjM2QwM2I4Njk1NjYzNjhhMWVhMjk1NzQ3NDJkMjRlOGZhYjg0YmNjOTg4OTVkMjY4YjEzNmZhNGEwMGUmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0In0.lSAgu595B_AgG_kgqw0GHhq_rXqm04pltrnlN0SXeGI)
> 
> > Interestingly, I have consistently found that Compatibility has one extra frame of latency in fullscreen mode compared to windowed mode, even in 4.4rc1. Maybe I'll open an inssue for this...
> 
> This is driver behavior, I'm pretty sure. OpenGL lets drivers do additional frame queueing if they want, but are expected to flush that queue when `glFinish` is called (which this PR does with low latency mode active). Could also change the number of queued/pre-rendered frames in the driver settings. The driver-level low-latency setting might also work.
> 
> Driver settings won't have any effect in D3D12 or Vulkan though. With those APIs, frame queueing is handled by the client application, so the only way to disable that queueing is with this PR.

Hah! I completely forgot to enable it. No idea why I missed this. I'll re-test!