@allenwp 2025-02-27
> This is driver behavior, I'm pretty sure. OpenGL lets drivers do additional frame queueing if they want, but are expected to flush that queue when `glFinish` is called (which this PR does with low latency mode active). Could also change the number of queued/pre-rendered frames in the driver settings. The driver-level low-latency setting might also work.
> 
> Driver settings won't have any effect in D3D12 or Vulkan though. With those APIs, frame queueing is handled by the client application, so the only way to disable that queueing is with this PR.

I re-did the test with low latency mode enabled and found:

- fullscreen Compatibility: reduced latency by around 2 frames
- windowed Compatibility: no impact on latency

This was a rough test where I did not use a high speed camera. That said, I have found that the [virtual cursor movement method](https://github.com/KeyboardDanni/godot-latency-tester) gives the same results as a [physical mouse movement](https://github.com/allenwp/in-out-latency-tester).

My results and video recordings are here: https://drive.google.com/drive/u/0/folders/1p4Ob0HHFqX5zzIbkXL13tqKTnh8EVy03
(Comparison-low-latency-mode.xlsx file has the summary)