@darksylinc 2025-03-28
The changes LGTM from a technical point. I marked something that could use editing in the Documentation
I was planning on taking a stab on this PR in April after I was done with TheForge UMA PR; my intention is to implement the automatic method; and hopefully introduce carefully-placed sleeps to minimize input latency.

When I said "LGTM from a technical point" is because I share the same issues @allenwp mentions:

1. It's a footgun.
2. **Psychological:** No matter what we do; users (both gamedev and players) **_will_** enable CPU\_GPU\_SYNC\_SEQUENTIAL, notice their FPS drops from 1000 FPS to 500 FPS and they **_will_** periodically open a bug report in Godot _insisting_ this is a bug that needs fixing. Because "losing 500 FPS is a lot, it's embarrassing". There will be nothing to fix: It's working as expected, and losing 500 FPS is not a lot. It's just 0.5ms. And even if you go from 50 FPS to 30 FPS (which is a lot, 13.33ms) it's still working as intended.
3. There's indeed a need to lower latency in the Vulkan renderer. Worst case we can easily see 4 frames of latency which is puke-inducing.
4. I'm constantly swinging between "we should only give the options 'AUTO' and 'PARALLEL'. If AUTO is broken then just tell the user to disable VSync" and "We still need to expose 'SEQUENTIAL' in case 'AUTO' is broken". Once AUTO is implemented, it should definitely be the default (that would fix a lot of issues with the Editor).

Except implementing AUTO (which hopefully should start somewhere in April), this is a social problem and not a technical one.

From technical point, changes LGTM (minus this [documentation change](https://github.com/godotengine/godot/pull/100031#discussion_r2019101111)).